health cards 2.0 refactor
