+++ .ruby-version
@@ -1,1 +1,1 @@
-2.7.3
+2.7.4
