+++ .tool-versions
@@ -1,2 +1,2 @@
-ruby 2.7.3
+ruby 2.7.4
 nodejs 14.16.0
