+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM ruby:2.7.3
+FROM ruby:2.7.4
 
 ENV INSTALL_PATH=/opt/health_cards/
 ENV RAILS_ENV=production
