+++ test_app/Gemfile
@@ -2,6 +2,6 @@
 
 source 'https://rubygems.org'
 
-ruby '2.7.3'
+ruby '2.7.4'
 
 gem 'health_cards', path: '..'
