+++ test_app/test_health_cards_without_rails
@@ -15,10 +15,10 @@ jws.verify
 ver = HealthCards::Verifier.new(keys: key.public_key)
 ver.verify(jws)
 
-card1 = HealthCards::HealthCard.from_jws(jws.to_s)
+card1 = HealthCards::HealthCard.new(jws.to_s)
 card1.to_json
 
-card2 = HealthCards::COVIDHealthCard.from_jws(jws.to_s)
+card2 = HealthCards::COVIDPayload.from_jws(jws.to_s)
 card2.to_json
 
 codes = HealthCards::QRCodes.from_jws(jws.to_s)
